LayoutTests/editing/pasteboard/merge-end-list-2-expected.txt

EDITING DELEGATE: shouldBeginEditingInDOMRange:range from 0 of DIV > BODY > HTML > #document to 1 of DIV > BODY > HTML > #document
EDITING DELEGATE: webViewDidBeginEditing:WebViewDidBeginEditingNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 7 of #text > DIV > LI > UL > DIV > BODY > HTML > #document to 7 of #text > DIV > LI > UL > DIV > BODY > HTML > #document toDOMRange:range from 3 of #text > DIV > LI > UL > DIV > BODY > HTML > #document to 3 of #text > DIV > LI > UL > DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
EDITING DELEGATE: webViewDidEndEditing:WebViewDidEndEditingNotification
This tests pasting text ('foo\nbar') into a list item. The last bit of content in the incoming fragment should be merged with the paragraph after the position being pasted into.

Before insertion:
| <ul>
|   <li>
|     <div>
|       "Item 1"

After insertion:
| <ul>
|   <li>
|     <div>
|       "Itemfoo"
|     <div>
|       "bar<#selection-caret> 1"
